+++ uplink/src/base/mod.rs
@@ -5,7 +5,7 @@ use log::{debug, trace};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-pub mod actions;
+pub mod middleware;
 pub mod mqtt;
 pub mod serializer;
